requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==1.3.0
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.54.0
django-simpleui==2023.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.4
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
soupsieve==2.4.1
sqlparse==0.4.4
tiktoken==0.3.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.9.2
